Refactor and DRY crud controllers
This is easy to follow and understand without having to mess around
with 2 different concerns, one overriding methods of the other one.
Also, a few lines less of code per controller is nice :-)
