+++ app/controllers/steps/applicant/contact_details_controller.rb
@@ -1,8 +1,6 @@
 module Steps
   module Applicant
     class ContactDetailsController < Steps::ApplicantStepController
-      include CrudStep
-
       def edit
         @form_object = ContactDetailsForm.build(
           current_record, c100_application: current_c100_application
@@ -15,12 +13,6 @@ module Steps
           record: current_record,
           as: :contact_details
         )
-      end
-
-      private
-
-      def record_collection
-        @_record_collection ||= current_c100_application.applicants
       end
     end
   end
