+++ app/controllers/steps/applicant_step_controller.rb
@@ -1,9 +1,17 @@
 module Steps
-  class ApplicantStepController < StepController
+  class ApplicantStepController < CrudStepController
     private
 
     def decision_tree_class
       C100App::ApplicantDecisionTree
+    end
+
+    def names_form_class
+      Applicant::NamesForm
+    end
+
+    def record_collection
+      @_record_collection ||= current_c100_application.applicants
     end
   end
 end
