+++ app/controllers/steps/children_step_controller.rb
@@ -1,9 +1,17 @@
 module Steps
-  class ChildrenStepController < StepController
+  class ChildrenStepController < CrudStepController
     private
 
     def decision_tree_class
       C100App::ChildrenDecisionTree
+    end
+
+    def names_form_class
+      Children::NamesForm
+    end
+
+    def record_collection
+      @_record_collection ||= current_c100_application.children
     end
   end
 end
