+++ app/controllers/steps/other_children_step_controller.rb
@@ -1,9 +1,17 @@
 module Steps
-  class OtherChildrenStepController < StepController
+  class OtherChildrenStepController < CrudStepController
     private
 
     def decision_tree_class
       C100App::OtherChildrenDecisionTree
+    end
+
+    def names_form_class
+      OtherChildren::NamesForm
+    end
+
+    def record_collection
+      @_record_collection ||= current_c100_application.other_children
     end
   end
 end
