+++ app/controllers/steps/respondent_step_controller.rb
@@ -1,9 +1,17 @@
 module Steps
-  class RespondentStepController < StepController
+  class RespondentStepController < CrudStepController
     private
 
     def decision_tree_class
       C100App::RespondentDecisionTree
+    end
+
+    def names_form_class
+      Respondent::NamesForm
+    end
+
+    def record_collection
+      @_record_collection ||= current_c100_application.respondents
     end
   end
 end
